Begin refactor of right sidebar
