+++ src/blocks.rs
@@ -1,4 +1,4 @@
 //! Smaller components like buttons, menus etc composed into larger widgets.
 
-pub mod right_side_bar;
+pub mod right_sidebar;
 pub mod workbench;
